[PARCOURS SECURISATION] Masque les mesures
... aux utilisateurs non connectés
